tests: Rename `helpers` folder to `test-helpers`
