+++ tests/acceptance/front-page-test.js
@@ -3,8 +3,8 @@ import { setupApplicationTest } from 'ember-qunit';
 import { currentURL, visit } from '@ember/test-helpers';
 import a11yAudit from 'ember-a11y-testing/test-support/audit';
 import axeConfig from '../axe-config';
-import { title } from '../helpers/dom';
-import setupMirage from '../helpers/setup-mirage';
+import { title } from '../test-helpers/dom';
+import setupMirage from '../test-helpers/setup-mirage';
 import { percySnapshot } from 'ember-percy';
 
 module('Acceptance | front page', function(hooks) {
