+++ tests/acceptance/search-test.js
@@ -4,8 +4,8 @@ import { fillIn, currentURL, triggerEvent, visit, blur } from '@ember/test-helpe
 import a11yAudit from 'ember-a11y-testing/test-support/audit';
 import { keyDown, keyPress } from 'ember-keyboard/test-support/test-helpers';
 import axeConfig from '../axe-config';
-import { title } from '../helpers/dom';
-import setupMirage from '../helpers/setup-mirage';
+import { title } from '../test-helpers/dom';
+import setupMirage from '../test-helpers/setup-mirage';
 import { percySnapshot } from 'ember-percy';
 
 module('Acceptance | search', function(hooks) {
